Reduce test suite duration
- use smaller datasets
- stop solvers after first solution discovery
- reduce time_limit

+++ src/discrete_optimization/generic_tools/callbacks/early_stoppers.py
@@ -23,7 +23,7 @@ class TimerStopper(Callback):
 
     """
 
-    def __init__(self, total_seconds: int, check_nb_steps: int = 1):
+    def __init__(self, total_seconds: float, check_nb_steps: int = 1):
         """
 
         Args:
